+++ src/components/auth/Signup.tsx
@@ -1,26 +1,19 @@
-import { useState } from "react";
-import { AiFillEyeInvisible, AiFillEye } from "react-icons/ai";
-import { signUp } from "../../api";
+import { Dispatch, FC, SetStateAction, useState } from "react";
+import AuthModal from "./AuthModal";
+import { TextField } from "@mui/material";
 import { useForm } from "react-hook-form";
+import { singUpformData } from "../../types/user";
+import { signUp } from "../../api";
 
-type Inputs = {
-  name: string;
-  userName: string;
-  email: string;
-  password: string;
-  passwordConfirm: string;
-};
+interface Props {
+  isOpen: boolean;
+  setIsOpen: Dispatch<SetStateAction<boolean>>;
+}
 
-const Signup = () => {
-  const inputStyle =
-    "w-[76%] h-[40px] mx-auto mb-5 block bg-[#DEDEDE] border-2 border-[#00000030] rounded pl-4 focus:outline-primary focus:bg-[#fff]";
+const Auth3: FC<Props> = ({ isOpen, setIsOpen }) => {
+  const { register, handleSubmit } = useForm<singUpformData>();
 
-  const [visiblePW, setVisiblePW] = useState<boolean>(false);
-  const [visiblePWC, setVisiblePWC] = useState<boolean>(false);
-
-  const { register, handleSubmit } = useForm<Inputs>();
-
-  const onSubmit = async (formData: Inputs) => {
+  const onSubmit = async (formData: singUpformData) => {
     try {
       const res = await signUp(formData);
       alert(res.data.message);
@@ -30,67 +23,42 @@ const Signup = () => {
   };
 
   return (
-    <form onSubmit={handleSubmit(onSubmit)}>
-      <input
+    <AuthModal
+      isOpen={isOpen}
+      setIsOpen={setIsOpen}
+      role="로그인"
+      submit={handleSubmit(onSubmit)}
+    >
+      <TextField
+        label="이름"
+        helperText="Rewind내에서 사용할 이름을 입력해주세요"
         type="text"
-        placeholder="이름"
-        className={inputStyle}
+        required
         {...register("name", { required: true })}
       />
-      <input
-        type="text"
-        placeholder="사용자 이름"
-        className={inputStyle}
-        {...register("userName", { required: true })}
-      />
-      <input
+      <TextField
+        label="계정"
+        helperText="이메일을 입력해주세요"
         type="email"
-        placeholder="사용자 이메일"
-        className={inputStyle}
+        required
         {...register("email", { required: true })}
       />
-      <input
-        type={visiblePW ? "text" : "password"}
-        placeholder="비밀번호"
-        className={inputStyle}
+      <TextField
+        label="비밀번호"
+        helperText="비밀번호를 입력해주세요"
+        type="password"
+        required
         {...register("password", { required: true })}
       />
-      {visiblePW ? (
-        <AiFillEye
-          className="block absolute right-[15%] top-[400px] cursor-pointer text-lg"
-          onClick={() => setVisiblePW(false)}
-        />
-      ) : (
-        <AiFillEyeInvisible
-          className="block absolute right-[15%] top-[400px] cursor-pointer text-lg"
-          onClick={() => setVisiblePW(true)}
-        />
-      )}
-      <input
-        type={visiblePWC ? "text" : "password"}
-        placeholder="비밀번호 확인"
-        className={inputStyle}
+      <TextField
+        label="비밀번호 확인"
+        helperText="비밀번호를 다시 입력해주세요"
+        type="password"
+        required
         {...register("passwordConfirm", { required: true })}
       />
-      {visiblePWC ? (
-        <AiFillEye
-          className="absolute right-[15%] top-[460px] cursor-pointer text-lg"
-          onClick={() => setVisiblePWC(false)}
-        />
-      ) : (
-        <AiFillEyeInvisible
-          className="absolute right-[15%] top-[460px] cursor-pointer text-lg"
-          onClick={() => setVisiblePWC(true)}
-        />
-      )}
-      <button
-        className="w-[76%] h-[40px]  block mx-auto rounded font-bold text-white bg-primary"
-        type="submit"
-      >
-        회원가입
-      </button>
-    </form>
+    </AuthModal>
   );
 };
 
-export default Signup;
+export default Auth3;
