+++ src/components/common/Main.tsx
@@ -1,14 +1,13 @@
 import { Outlet } from "react-router-dom";
-import Header from "./Header";
+import Header from "../header/Header";
+import Container from "@mui/material/Container";
 
 const Main = () => {
   return (
-    <div className="bg-[#f2f2f2] h-[100vh]">
-      <div className="w-[100vw] mx-auto max-w-[1400px] px-3">
-        <Header />
-        <Outlet />
-      </div>
-    </div>
+    <Container maxWidth="xl">
+      <Header />
+      {/* <Outlet /> */}
+    </Container>
   );
 };
