+++ src/components/fallback/Loading.tsx
@@ -1,5 +1,11 @@
+import CircularProgress from "@mui/material/CircularProgress";
+
 const Loading = () => {
-  return <div>Loading</div>;
+  return (
+    <div className="h-[100vh] flex items-center justify-center">
+      <CircularProgress />
+    </div>
+  );
 };
 
 export default Loading;
